% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{pcpca_mle}
\alias{pcpca_mle}
\title{PCPCA MLE}
\usage{
pcpca_mle(X, Y, gamma, d = 1, verbose = TRUE)
}
\arguments{
\item{X}{matrix of the FOREGROUND data}

\item{Y}{matrix of the BACKGROUND data}

\item{gamma}{PCPCA tuning parameter}

\item{d}{latent dimension to project to}

\item{verbose}{verbosity switch}
}
\value{
a list with two elements: sigma2_mle and W_mle
}
\description{
PCPCA MLE
}
