% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{pcpca_mcmc}
\alias{pcpca_mcmc}
\title{EZPZ MCMC PCPCA}
\usage{
pcpca_mcmc(X, Y, gamma, d = 1, verbose = TRUE, mle_est, init = NULL)
}
\arguments{
\item{X}{matrix of the FOREGROUND data}

\item{Y}{matrix of the BACKGROUND data}

\item{gamma}{PCPCA tuning parameter}

\item{d}{latent dimension to project to}

\item{verbose}{verbosity switch}

\item{mle_est}{description}

\item{init}{initialization list or function}
}
\description{
EZPZ MCMC PCPCA
}
\details{
If set, \code{init} is passed to cmdstanr::sample() as is.

For the sake of identifiability, the largest element of the first axis is
constrained to be positive, and the cross product of the first two
contrastive axes is positive for d = 2.
}
